feat: implement RegisterForm component for user registration
